+++ src/lib/config/app-paths.ts
@@ -1,5 +1,6 @@
 export const AppPaths = {
   home: '/',
+  dashboard: '/dashboard',
   auth: {
     login: '/auth/login',
   },